Stage 4: enforce module boundary, remove dead code, document the bridge
- tests/test_boundaries.py: game-side modules must import no rendering/GUI
  libraries (subprocess import-purity gate, complements the stage-3
  headless functional test)
- graphics.py: remove unused LineOfSightFilter1, LineOfSightFilter,
  SightRenderer, LOSTextureRenderer, CPUShadowRenderer stub and
  MemoryRenderer (366 lines); drop now-unused Varying/PolarTransform
  imports. Sprites alias and CharAtlas stay (ui.py uses them).
- scene.py: line_of_sight normalized to float32 once at the provider
  boundary; todo note on the pre-existing missing take/items_at features
- ARCHITECTURE.md: reference for the game/renderer bridge contract
  (layers, versions, visibility provider, frame protocol, messages) and
  what a second backend implements

Verification: 26 tests pass; screenshot regression vs the pre-refactor
baseline bit-identical; game runs live with no traceback.

+++ carriage_return/scene.py
@@ -99,6 +99,8 @@ class Scene(Entity):
         self.messages(message=message)
 
     def request_player_action(self, action):
+        # todo: items_at() and Player.take() don't exist yet (predates the
+        # renderer split); 'take' will raise until inventory pickup is built
         if action == 'take':
             items = self.items_at(self.player.location.slot)
             if len(items) == 0:
@@ -125,7 +127,7 @@ class Scene(Entity):
         """
         # render new line of sight
         if self._need_los_update:
-            self.line_of_sight = self.player.line_of_sight()
+            self.line_of_sight = self.player.line_of_sight().astype('float32', copy=False)
             self._need_los_update = False
 
         # calculate lighting
